requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
streamlit>=1.33.0
pydantic>=2.6.0
torch>=2.2.0
transformers>=4.39.0
peft>=0.10.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
openai>=1.30.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
datasets>=2.19.0
